FastAPI/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
pydantic-settings>=2.0.3
email-validator>=2.1.0
